moved tests out of src directory
