+++ tests/regression/thread-locking.test.js
